+++ src/main/java/carrace/domain/Movable.java
@@ -1,4 +1,4 @@
-package carrace;
+package carrace.domain;
 
 @FunctionalInterface
 public interface Movable {
