+++ src/main/java/carrace/domain/RandomMovable.java
@@ -1,4 +1,4 @@
-package carrace;
+package carrace.domain;
 
 import java.util.Random;
